Ruff stops complaining about annotations, more reorganization
