+++ pyproject.toml
@@ -21,7 +21,7 @@ select = [
     "I",   # isort
     "UP",  # pyupgrade
     "G",   # logging
-    "FA",  # future annotations
+    # "FA",  # future annotations
     "PIE", # misc
 ]
