+++ src/neutropenia_redcap/redcap/forms/germline/generic.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class GermlineForm:
+    mrn: int
