+++ src/neutropenia_redcap/redcap/forms/germline/scnir.py
@@ -1,6 +1,15 @@
-from collections.abc import Sequence
+from __future__ import annotations
+
+from collections.abc import Collection, Iterable, Sequence
+from dataclasses import dataclass
 from functools import partial
 from itertools import chain
+
+import polars as pl
+from more_itertools import padded
+from variants.germline.scnir import SCNIRGermlineGeneMention
+
+from .generic import GermlineForm
 
 MINIMUM_SCNIR_GERMLINES = 1
 MAXIMUM_SCNIR_GERMLINES = 3
@@ -54,3 +63,32 @@ SCNIR_GERMLINE_COLUMNS = list(
         ),
     )
 )
+
+SCNIR_GERMLINE_SCHEMA = [
+    (column_name, pl.String) for column_name in SCNIR_GERMLINE_COLUMNS
+]
+
+
+@dataclass
+class SCNIRGermlineForm(GermlineForm):
+    gene_mentions: Collection[SCNIRGermlineGeneMention]
+
+    def to_row(self) -> Iterable[str | bool | None]:
+        # patient_id
+        yield self.mrn
+        # sum_germ, 1 == "Yes"
+        yield 1
+        # sum_germ_num_gen
+        yield min(len(self.gene_mentions), MAXIMUM_SCNIR_GERMLINES)
+        for germline in padded(
+            self.gene_mentions, n=MAXIMUM_SCNIR_GERMLINES, fillvalue=None
+        ):
+            yield from (
+                SCNIRGermlineGeneMention.blank_row_fragment()
+                if germline is None
+                else germline.to_row_fragment()
+            )
+
+    def to_data_frame(self) -> pl.DataFrame:
+        data = [list(self.to_row())]
+        return pl.DataFrame(data=data, schema=SCNIR_GERMLINE_SCHEMA, orient="row")
