+++ src/neutropenia_redcap/variants/germline/scnir.py
@@ -1,247 +1,15 @@
 from __future__ import annotations
 
-import datetime
-import logging
 from collections.abc import Collection, Iterable
 from dataclasses import dataclass, field
-from datetime import date
-from functools import cache
-from operator import attrgetter, is_not_none
-from typing import ClassVar
 
-import polars as pl
-from more_itertools import padded, partition
-
-from .redcap_import.germline.scnir import (
+from more_itertools import padded
+from redcap.forms.germline.scnir import (
     MAXIMUM_SCNIR_GERMLINE_VARIANTS,
-    MAXIMUM_SCNIR_GERMLINES,
     MINIMUM_SCNIR_GERMLINE_VARIANTS,
-    SCNIR_GERMLINE_COLUMNS,
 )
 
-logger = logging.getLogger(__name__)
-
-logging.basicConfig(
-    format="%(asctime)s - %(levelname)s - %(name)s -   %(message)s",
-    datefmt="%m/%d/%Y %H:%M:%S",
-    level=logging.INFO,
-)
-
-SCNIR_GERMLINE_SCHEMA = [
-    (column_name, pl.String) for column_name in SCNIR_GERMLINE_COLUMNS
-]
-
-VARIANT_TYPES = [
-    "Pathogenic",
-    "Likely Pathogenic",
-    "Benign",
-    "Likely Benign",
-    "Uncertain",
-]
-
-
-@cache
-def map_variant_type(variant_type: str | None) -> int | None:
-    if variant_type is None:
-        return None
-    normalized_variant_type = " ".join(variant_type.lower().split())
-    is_likely = "likely" in normalized_variant_type
-    is_benign = "benign" in normalized_variant_type
-    is_pathogenic = "patho" in normalized_variant_type
-    is_uncertain = (
-        "uncertain" in normalized_variant_type
-        or "vus" in normalized_variant_type
-        or "unknown significance" in normalized_variant_type
-    )
-    match is_likely, is_benign, is_pathogenic, is_uncertain:
-        # Unmodified pathogenic
-        case False, False, True, False:
-            return 1
-        # Likely pathogenic
-        case True, False, True, False:
-            return 2
-        # Unmodified benign
-        case False, True, False, False:
-            return 3
-        # Likely benign
-        case True, True, False, False:
-            return 4
-        # Unmodified uncertain
-        case False, False, False, True:
-            return 5
-    logger.warning("Cannot currently map: %s", variant_type)
-    return None
-
-
-@dataclass(eq=True, frozen=True)
-class TextSource:
-    filename: str
-    section: str
-    sentence: str
-    file_date: datetime.date | None
-
-
-@dataclass(eq=True, frozen=True)
-class SomaticVariant:
-    gene: str
-    syntax_p: str | None
-    syntax_n: str | None
-    variant_type: str | None
-    vaf: str | None
-    heterozygous: (
-        bool | None
-    )  # True for is heterozygous, False for definitely isn't, None for unknown
-    text_sources: Collection[TextSource] = field(compare=False)
-    specimen_collection_dates: Collection[date] = field(compare=False)
-    sample_sources: Collection[str] = field(compare=False)
-
-    # TODO a lot of the potential methods might be coordinated with the
-    # germline variant code
-
-
-@dataclass(eq=True, frozen=True)
-class GermlineVariant:
-    gene: str
-    syntax_p: str | None
-    syntax_n: str | None
-    variant_type: str | None
-    vaf: str | None
-    heterozygous: (
-        bool | None
-    )  # True for is heterozygous, False for definitely isn't, None for unknown
-    text_sources: Collection[TextSource] = field(compare=False)
-    specimen_collection_dates: Collection[date] = field(compare=False)
-    sample_sources: Collection[str] = field(compare=False)
-    # protein syntax, nucleotide syntax, variant type, comment
-    total_variant_attrs: ClassVar[int] = 4
-
-    # Another weird thing is I can't find the field where the specimen collection date
-    # would go
-    def to_row_fragment(self, blank: bool = False) -> Iterable[str | bool | None]:
-        if blank:
-            yield from GermlineVariant.blank_row_fragment(
-                total_variant_attrs=GermlineVariant.total_variant_attrs
-            )
-        variant_type, source = self.select_variant_type()
-        # sum_germ_var{variant_index}_cdna_{germline_index}
-        yield self.syntax_n
-        # sum_germ_var{variant_index}_pro_{germline_index}
-        yield self.syntax_p
-        # sum_germ_var{variant_index}_acmg_{germline_index}
-        yield variant_type
-        # sum_germ_var{variant_index}_comment_{germline_index}
-        yield self.build_comment(variant_type=variant_type, source=source)
-
-    # Heterozygosity in "Clinical and Research Sequencing Summary Form Comments"
-    # or variant level comments
-    def build_comment(self, variant_type: int | None, source: str) -> str:
-        mention_summary = self.build_mention_summary(
-            variant_type=variant_type, source=source
-        )
-        source_guide = self.build_source_guide()
-        return mention_summary + source_guide
-
-    def build_mention_summary(self, variant_type: int | None, source: str) -> str:
-        normalized_syntax_n = (
-            self.syntax_n.lower() if self.syntax_n is not None else "None found"
-        )
-        normalized_syntax_p = (
-            self.syntax_p.lower() if self.syntax_p is not None else "None found"
-        )
-        match self.heterozygous:
-            case None:
-                heterozygous = "Unknown"
-            case False:
-                heterozygous = "No"
-            case True:
-                heterozygous = "Yes"
-            case _:
-                raise ValueError(
-                    f"Improper value for heterozygous field of SCNIR variant {self.heterozygous} - should be a boolean or None"
-                )
-        normalized_variant_type = (
-            VARIANT_TYPES[variant_type - 1] if variant_type is not None else "Unknown"
-        )
-        return (
-            "Mention Summary:\n"
-            f"Gene: {self.gene.upper()} Nucleotide Syntax: {normalized_syntax_n} Protein Syntax: {normalized_syntax_p} Variant Type (Parsed from {source.title()}): {normalized_variant_type} Heterozygous: {heterozygous}\n\n"
-        )
-
-    def select_variant_type(self) -> tuple[int | None, str]:
-        model_parsed_variant_type = map_variant_type(self.variant_type)
-        if model_parsed_variant_type is not None:
-            return model_parsed_variant_type, "sentence"
-        return self.select_variant_type_from_sources()
-
-    def select_variant_type_from_sources(self) -> tuple[int | None, str]:
-        sections = list(map(attrgetter("section"), self.text_sources))
-        mapped_sections = list(
-            filter(is_not_none, map(map_variant_type, sections)),
-        )
-        variant_type = None
-        match len(mapped_sections):
-            case 0:
-                variant_type = None
-            case 1:
-                variant_type = mapped_sections[0]
-            case _:
-                if all(vtype == mapped_sections[0] for vtype in mapped_sections):
-                    variant_type = mapped_sections[0]
-                else:
-                    logger.error(
-                        "Variant type inconsistencies with sections, using None: %s",
-                        ", ".join(sections),
-                    )
-                    variant_type = None
-        return variant_type, "sections"
-
-    def build_source_guide(self) -> str:
-        match len(self.text_sources):
-            case 0:
-                raise ValueError("Mention has no text sources")
-            case 1:
-                return GermlineVariant.build_single_source_guide(
-                    text_source=next(iter(self.text_sources))
-                )
-            case _:
-                return GermlineVariant.build_multi_source_guide(
-                    text_sources=self.text_sources
-                )
-
-    @staticmethod
-    def build_single_source_guide(text_source: TextSource) -> str:
-        return f"Found in sentence:\n{text_source.sentence}\nFrom file {text_source.filename} - section '{text_source.section}'"
-
-    @staticmethod
-    def build_multi_source_guide(text_sources: Collection[TextSource]) -> str:
-        no_date_info, has_date_info = partition(
-            lambda source: source.file_date is not None, text_sources
-        )
-        sorted_sources = sorted(has_date_info, key=attrgetter("file_date"))
-        others = (
-            f"From file {text_source.filename} - section '{text_source.section}'"
-            for text_source in sorted_sources[1:]
-        )
-        proper = f"Multiple sources found, showing earliest:\n{GermlineVariant.build_single_source_guide(sorted_sources[0])}\n\nOthers least to most recent:\n{'\n'.join(others)}\n"
-        no_date_info = list(no_date_info)
-        if len(no_date_info) > 0:
-            no_date_info_mentions = "\n".join(
-                f"From file {text_source.filename} - section '{text_source.section}'"
-                for text_source in no_date_info
-            )
-            return f"{proper}Others with no date information:\n{no_date_info_mentions}"
-        return proper
-
-    @staticmethod
-    def blank_row_fragment(total_variant_attrs: int) -> Iterable[None]:
-        for _ in range(total_variant_attrs):
-            yield None
-
-
-@dataclass(eq=True, frozen=True)
-class GermlineGeneMention:
-    gene: str
-    variants: Collection[GermlineVariant] = field(compare=False)
+from .generic import GermlineVariant
 
 
 @dataclass(eq=True, frozen=True)
@@ -277,33 +45,3 @@ class SCNIRGermlineGeneMention:
             yield from GermlineVariant.blank_row_fragment(
                 total_variant_attrs=GermlineVariant.total_variant_attrs
             )
-
-
-@dataclass
-class GermlineForm:
-    mrn: int
-
-
-@dataclass
-class SCNIRGermlineForm(GermlineForm):
-    gene_mentions: Collection[SCNIRGermlineGeneMention]
-
-    def to_row(self) -> Iterable[str | bool | None]:
-        # patient_id
-        yield self.mrn
-        # sum_germ, 1 == "Yes"
-        yield 1
-        # sum_germ_num_gen
-        yield min(len(self.gene_mentions), MAXIMUM_SCNIR_GERMLINES)
-        for germline in padded(
-            self.gene_mentions, n=MAXIMUM_SCNIR_GERMLINES, fillvalue=None
-        ):
-            yield from (
-                SCNIRGermlineGeneMention.blank_row_fragment()
-                if germline is None
-                else germline.to_row_fragment()
-            )
-
-    def to_data_frame(self) -> pl.DataFrame:
-        data = [list(self.to_row())]
-        return pl.DataFrame(data=data, schema=SCNIR_GERMLINE_SCHEMA, orient="row")
