+++ src/neutropenia_redcap/variants/sources.py
@@ -1,0 +1,10 @@
+import datetime
+from dataclasses import dataclass
+
+
+@dataclass(eq=True, frozen=True)
+class TextSource:
+    filename: str
+    section: str
+    sentence: str
+    file_date: datetime.date | None
